requirements.txt
chainlit
streamlit
openai
langchain
langchain-community
langchain-openai
python-dotenv
astrapy